pyproject.toml
[project]
name = "jtr-generator"
version = "0.1.0"
description = "日本の履歴書（JIS規格準拠）をPDF形式で生成するツール"
authors = [
    { name = "jtr-generator contributors" }
]
readme = "jtr-generator/references/README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "reportlab>=4.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.20.0",
    "mistune>=3.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["jtr-generator"]

# Ruff設定
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["jtr-generator", "tests", "tools"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["jtr-generator", "jtr"]

# mypy設定
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = "reportlab.*"
ignore_missing_imports = true

# pytest設定
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
]
# pytest-spec設定（docstringを優先表示）
spec_test_format = "{result} {docstring_summary}"

# pytest-cov設定
[tool.coverage.run]
source = ["jtr-generator", "tools"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 80

# poethepoet タスクランナー設定
[tool.poe.tasks]
test = "pytest tests/ --spec"
test-cov = "pytest tests/ --cov=jtr-generator --cov=tools --cov-report=term-missing --cov-report=xml --spec"
test-watch = "pytest tests/ -f --spec"
lint = "ruff check jtr-generator/ tests/ tools/"
lint-fix = "ruff check --fix jtr-generator/ tests/ tools/"
format = "ruff format jtr-generator/ tests/ tools/"
format-check = "ruff format --check jtr-generator/ tests/ tools/"
typecheck = "mypy jtr-generator/"
check-coverage = "python tools/check_coverage_per_file.py"
clean-outputs = "rm -rf outputs/*.pdf outputs/*.jpg outputs/debug/"

# check-allは順次実行（test-cov完了後にcheck-coverageを実行）
[tool.poe.tasks.check-all]
help = "Run all quality checks"
sequence = [
    "lint",
    "format-check",
    "typecheck",
    "test-cov",
    "check-coverage"
]

# Agent Skillsパッケージビルド
[tool.poe.tasks.build-skill]
help = "Build Agent Skills package (jtr-generator.zip)"
sequence = [
    { cmd = "cd jtr-generator && zip -r ../build/jtr-generator.zip . -x \"*/__pycache__/*\"" },
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "poethepoet>=0.39.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-spec>=4.0.0",
    "ruff>=0.14.10",
    "types-jsonschema>=4.25.1.20251009",
    "types-pyyaml>=6.0.12.20250915",
    "pypdf>=6.5.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    # 画像処理・視覚的検証用
    "pdf2image>=1.17.0",
    "scikit-image>=0.24.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
]
